#5186 MView edit + comment fixes
Former-commit-id: 766989d53117a030eeb290510afc38048ff2659d

+++ plugins/org.jkiss.dbeaver.ext.oracle/src/org/jkiss/dbeaver/ext/oracle/edit/OracleMaterializedViewManager.java
@@ -19,13 +19,14 @@ package org.jkiss.dbeaver.ext.oracle.edit;
 
 import org.jkiss.code.Nullable;
 import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.ext.oracle.model.OracleMaterializedView;
+import org.jkiss.dbeaver.ext.oracle.model.OracleSchema;
 import org.jkiss.dbeaver.model.DBPDataSource;
 import org.jkiss.dbeaver.model.DBPEvaluationContext;
 import org.jkiss.dbeaver.model.DBPScriptObject;
-import org.jkiss.dbeaver.model.edit.DBEPersistAction;
-import org.jkiss.dbeaver.ext.oracle.model.OracleMaterializedView;
-import org.jkiss.dbeaver.ext.oracle.model.OracleSchema;
 import org.jkiss.dbeaver.model.edit.DBECommandContext;
+import org.jkiss.dbeaver.model.edit.DBEPersistAction;
+import org.jkiss.dbeaver.model.edit.prop.DBECommandComposite;
 import org.jkiss.dbeaver.model.impl.DBSObjectCache;
 import org.jkiss.dbeaver.model.impl.edit.SQLDatabasePersistAction;
 import org.jkiss.dbeaver.model.impl.sql.edit.SQLObjectEditor;
@@ -41,10 +42,6 @@ import java.util.Map;
  * OracleMaterializedViewManager
  */
 public class OracleMaterializedViewManager extends SQLObjectEditor<OracleMaterializedView, OracleSchema> {
-    @Override
-    public boolean canEditObject(OracleMaterializedView object) {
-        return false;
-    }
 
     @Override
     public long getMakerOptions(DBPDataSource dataSource)
@@ -82,13 +79,13 @@ public class OracleMaterializedViewManager extends SQLObjectEditor<OracleMateria
     @Override
     protected void addObjectCreateActions(DBRProgressMonitor monitor, List<DBEPersistAction> actions, ObjectCreateCommand command, Map<String, Object> options)
     {
-        createOrReplaceViewQuery(actions, command.getObject());
+        createOrReplaceViewQuery(actions, command);
     }
 
     @Override
     protected void addObjectModifyActions(DBRProgressMonitor monitor, List<DBEPersistAction> actionList, ObjectChangeCommand command, Map<String, Object> options)
     {
-        //createOrReplaceViewQuery(actionList, command.getObject());
+        createOrReplaceViewQuery(actionList, command);
     }
 
     @Override
@@ -99,18 +96,29 @@ public class OracleMaterializedViewManager extends SQLObjectEditor<OracleMateria
         );
     }
 
-    private void createOrReplaceViewQuery(List<DBEPersistAction> actions, OracleMaterializedView view)
+    private void createOrReplaceViewQuery(List<DBEPersistAction> actions, DBECommandComposite<OracleMaterializedView, PropertyHandler> command)
     {
+        OracleMaterializedView view = command.getObject();
+
         StringBuilder decl = new StringBuilder(200);
         final String lineSeparator = GeneralUtils.getDefaultLineSeparator();
-        decl.append("CREATE MATERIALIZED VIEW ").append(view.getFullyQualifiedName(DBPEvaluationContext.DDL)).append(lineSeparator) //$NON-NLS-1$
-            .append("AS ").append(view.getObjectDefinitionText(null, DBPScriptObject.EMPTY_OPTIONS)); //$NON-NLS-1$
-        if (view.isPersisted()) {
+        boolean hasComment = command.getProperty("comment") != null;
+        if (!hasComment || command.getProperties().size() > 1) {
+            decl.append("CREATE MATERIALIZED VIEW ").append(view.getFullyQualifiedName(DBPEvaluationContext.DDL)).append(lineSeparator) //$NON-NLS-1$
+                .append("AS ").append(view.getObjectDefinitionText(null, DBPScriptObject.EMPTY_OPTIONS)); //$NON-NLS-1$
+            if (view.isPersisted()) {
+                actions.add(
+                    new SQLDatabasePersistAction("Drop view", "DROP MATERIALIZED VIEW " + view.getFullyQualifiedName(DBPEvaluationContext.DDL))); //$NON-NLS-2$
+            }
             actions.add(
-                new SQLDatabasePersistAction("Drop view", "DROP MATERIALIZED VIEW " + view.getFullyQualifiedName(DBPEvaluationContext.DDL))); //$NON-NLS-2$
+                new SQLDatabasePersistAction("Create view", decl.toString()));
         }
-        actions.add(
-            new SQLDatabasePersistAction("Create view", decl.toString()));
+        if (hasComment) {
+            actions.add(new SQLDatabasePersistAction(
+                "Comment table",
+                "COMMENT ON MATERIALIZED VIEW " + view.getFullyQualifiedName(DBPEvaluationContext.DDL) +
+                    " IS '" + view.getComment() + "'"));
+        }
     }
 
 }
